% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInfo.R
\name{extract_info}
\alias{extract_info}
\title{Extract information from Info.csv file}
\usage{
extract_info(filepath)
}
\arguments{
\item{filepath}{}
}
\description{
Extract information from Info.csv file
}
