% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_distances.R
\name{jac_distance_mat}
\alias{jac_distance_mat}
\title{Return matrix of jaccard distances of allele frequency matrix.
Samples in columns}
\usage{
jac_distance_mat(gene_AF_mat)
}
\arguments{
\item{gene_AF_mat}{}
}
\description{
Return matrix of jaccard distances of allele frequency matrix.
Samples in columns
}
