% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_similiarties.R
\name{jacc_distance}
\alias{jacc_distance}
\title{Return Jaccard distance for two vectors}
\usage{
jacc_distance(vec1, vec2)
}
\arguments{
\item{vec2}{}
}
\description{
Return Jaccard distance for two vectors
}
