% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_similiarties.R
\name{overlap_coef_mat}
\alias{overlap_coef_mat}
\title{Overlap coefficient matrix}
\usage{
overlap_coef_mat(gene_AF_mat)
}
\arguments{
\item{gene_AF_mat}{}
}
\description{
Overlap coefficient matrix
}
