% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_similiarties.R
\name{overlap_distance}
\alias{overlap_distance}
\title{Calculates overlap coefficient between two vectors}
\usage{
overlap_distance(vec1, vec2)
}
\arguments{
\item{vec2}{}
}
\description{
Calculates overlap coefficient between two vectors
}
