% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_activityList.R
\name{readin_actList_entry}
\alias{readin_actList_entry}
\title{Function to read in activityList_Entry from 'combined_output.xlsx' file}
\usage{
readin_actList_entry(filepath)
}
\arguments{
\item{filepath}{}
}
\description{
Function to read in activityList_Entry from 'combined_output.xlsx' file
}
