% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prep_snv.R
\name{readin_prepfiles}
\alias{readin_prepfiles}
\title{Function to read in watchdog prep_snv.txt files}
\usage{
readin_prepfiles(filepath)
}
\arguments{
\item{filepath}{}
}
\description{
Function to read in watchdog prep_snv.txt files
}
