% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Snvindel.R
\name{readin_snvindel}
\alias{readin_snvindel}
\title{Function to read in and file Snvindel.tsv files}
\usage{
readin_snvindel(filepath)
}
\arguments{
\item{filepath}{}
}
\description{
Function to read in and file Snvindel.tsv files
}
