% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortenName.R
\name{shorten_name}
\alias{shorten_name}
\title{Remove long absolute filepaths and GNXS UID}
\usage{
shorten_name(filestring)
}
\arguments{
\item{filestring}{}
}
\description{
Remove long absolute filepaths and GNXS UID
}
